Prevent the div.comments-body being overwritten
The display order form line (87) was making $output equal to
displayOrderForm(), accidentally overwriting the div.comments-body from 2
lines above. It should have been, and is now, appending displayOrderForm()
to the $output string.
This was causing there to be an extra </div>, which could be causing
problems (allthough luckily some browsers will have been able to cope)

Change-Id: I0e62491c5912422ca1a93b60805b03923472f048