Update patch set 1
Patch Set 1: Code-Review-1

The code and the approach here are OK; the reason why I'm giving this a -1 is a tad bit more complicated.

Long story short, this was a problem in 2010 (see r77271) and this still is a problem. Solving it in core Database class(es), as per the comment on line 1013 in CommentClass.php, would be better; Comments is only one of the many social tools that uses UNIX_TIMESTAMP.

But the *best* approach I'd like to see is to convert the SQL schemas to use binary(14) instead of datetime to store the timestamp, as the standard MediaWiki convention is. For Comments, this'd affect Comments.Comment_Date and Comments_Vote.Comment_Vote_Date fields. Then the code which deals with timestamps would need to be changed to use wfTimestamp(Now)()/$db->timestamp().

But that's the *easy* part. I'd have done this a while ago were it not for the fact that a migration script needs to be written (for any and all affected extensions), because users most certainly don't want to lose their valuable data -- or even parts of it -- due to some silly backend change that matters only to us developers!
Once such a migration script would be written, it'd need to be copied over to other social tools and adapted as needed in order to properly get rid of UNIX_TIMESTAMP for once and for all. Finally, the user-facing documentation on MW.org needs to be updated so that users looking to upgrade their social tools are aware that they need to run a migration script.

Patch-set: 1
Reviewer: Jack Phoenix <243@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1