Update patch set 1
Patch Set 1:

The documentation pages for social tools all use the "MW stable branch number" template for the minimum required version. Generally speaking this is true, although for the time being ShoutWiki is running MediaWiki 1.26, and where [[Template:MW stable branch version number]] and the version that ShoutWiki is running differ, ShoutWiki's version takes precedence (as it's usually lower). As for why this is the case for ShoutWiki, see https://blog.shoutwiki.com/2016/08/missingtype-and-mistakes/ for some background info.

Patch-set: 1