[WIP] Version 4.3: redesigned per Samantha's mockup
Needs quite a lot of PHP love and some CSS love.

There are some hardcoded English strings in this initial attempt which
will be fixed before this is merged.

CSS/LESS will also need cleanup, right now there are two files -- one of
each --, the old .css file is needed for the proper formatting of the
listing of pre-existing comments whereas the new LESS file is needed for
the comment submission box.

What works:
* submitting a comment
* replying to a comment
* previewing a reply (new!)

Needs more testing:
* how does the form look when the user doesn't have the 'comment' user right?
* and how does it behave when commenting is restricted to a subset of users _or_ the current user is blocked?

Known to be broken:
* character count isn't reset back to 0/1500 after submitting a comment
* clicking on the "cancel" button makes the "Reply to <username>" text (#replyto) disappear only once
* a lot of other stuff I didn't even think of testing, probably

Change-Id: Id23c85dae04678bc17c9c77c7267d5defb7be01a