Allow comments to be previewed prior to submission
Adds a "show preview" button next to the "post comment" button.
Pressing on the button sends an AJAX request to the API, which turns user input into properly parsed wikitext
and injects it between the comment textarea and the buttons.
While previewing, the "show preview" button is changed to "continue editing" and the textarea is grayed out.
If the user chooses "continue editing", the preview will be dismissed and the textarea becomes editable again,
allowing them to change their comment more before sending it.

Bug: T156055
Change-Id: I5d0118de21a4c61ff368fe77b9f7d7392cc08697