Update patch set 3
Patch Set 3: Code-Review+2

This is *a* solution but not exactly *the* best solution IMO, but that's not your fault. (Then again I'm not sure if using Reflection would be much cleaner either.) I'm not sure what good it serves to be locking down previously public Echo APIs like that and encourage people to duplicate code instead of something sane, but I guess that's what we have to live with, eh.

Patch-set: 3
Reviewer: Jack Phoenix <243@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2