Update patch set 3
Patch Set 3:

(1 comment)

> previously public Echo APIs

It was only public for a short time when it was made public just to be able to reuse it.

> instead of something sane

Sorry, but the complex dependency graph created by calling foreign code is not more "sane". Echo is not a library. It's another, mostly unrelated extension.

There is a lot written about why code reuse can be bad. I suggest to start with this post by a fellow college: https://www.entropywins.wtf/blog/2017/09/06/the-fallacy-of-dry/

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>