Actor migration: Fix order of db patches. Drop index before column
The upgrade script has the order of database patches wrong. You can't delete
a column that's part of a key in MySQL, that in this case gives a cryptic
error message: Key column 'Comment_Vote_Username' doesn't exist in table.

Moving the deletion of the key before the deletion of the column fixes this
issue.

Bug: T251713
Change-Id: I38afc1384383c7484eede126cb41500c92e93184