Update patch set 5
Patch Set 5:

(4 comments)

Always great to have more performance improvements, they're sorely needed! (And especially so in things like Comments which explicitly *disable* the (parser) cache.)

Now, one thing that _does_ concern me here (besides some random, rather insignificant inline nitpicks) is the fact that this new approach seems to be reliant on JavaScript. In my older, currently unmerged patch (https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Comments/+/643292) I made Comments largely function without JavaScript, as it is my belief as a software developer that most things which currently "require" JS...actually don't, and with many a [[mw:social tools]] this has been the case: a lot of JS dependencies have been just cases of the original authors being somewhat lazy. And it's understandable, because even back in 2006-2007 JavaScript was more likely to be enabled than disabled on the end-user's browser.
This patch, on the other hand, takes us to an entirely opposite direction, and from the viewpoint of a no-JS user, it's a regression over the current behavior, as (to my understanding) once this patch is merged, merely *viewing* comments requires JavaScript being enabled; personally the word that comes to mind is "obnoxious", though I'm aware that there might not be a perfect middle ground here when it comes to no-JS usability vs. performance.

All this said, while I may not like it, it seems clear to me that in weighing performance against no-JS usability, performance matters more to a larger group of users and use cases, so I'm not opposed to merging this, though I'd love if we could at least have a <noscript>ed message shown in the <comments /> callback so that no-JS users aren't shown an obviously incorrect "Loading comments.." message if and when loading comments requires JS to be enabled.

Patch-set: 5