Adds ability for users to edit own comments
'commentadmin' right grants an ability to edit anyone's comments.
Revisions are not saved, but the fact that a comment was edited
is saved in the comments log.

Also noticed that if EchoMentionOnChanges is false, Comment::add
hook would be skipped.  Fixed that and refactored Echo notifications
into their own function.

Bug: T146589
Change-Id: I71953f909a1044788f8ab9e22517c2cc1f3f1b8b