Scroll to the proper position when window.location.hash is set, e.g. by Comments' permalink feature
Not actually tested with paginated comments, but at least if and when a comment is on the first page and window.location.hash is set to #comment-<that comment's ID>, this code makes it jump to the proper position.

Follow-up to 009ccd6907d4b8102da7433dd5be172ca8ec8a5d.

Bug: T295567
Change-Id: I3088ddb4e9705be6ce78dfcda3d9d06f4ff4f08e