AbuseFilter interoperability
Borrows a lot of code from ArticleFeedbackv5.
Tested on a 1.35 wiki with the REL1_35 version of AbuseFilter, but should hopefully work even with newer versions of MW/AF.

AF filters can be set to affect comments only.
For comment submissions, the comment text is held as the AF variable "new_wikitext", so filters can use that var to match text.

Also fixes a tiny i18n issue with other, previously existing anti-spam checks: before this patch the properly i18n'd text would be wrapped in < > because a string, instead of a Message object, was being passed to dieWithError().
Now a Message object is passed instead, which fixes that issue.

Bug: T301083
Change-Id: Id39fd2de53a4af9dee1c8ff5d4e157542c66ab46