Hide tables of content in final comments and [edit source] links in comment preview
*Not* hiding the ToC in preview mode because I think it _can_ be useful for longer comments, because sometimes people (are) overly verbose.

This is a very crude hack since using ParserOutput properly in Comment#getText just does not seem to work the way it should, despite that since MW 1.35 Parser#setOptions *is* a thing.

Though the bug report speaks (only) of [edit] links, I observed that at least in MW 1.39 the only [edit] links that get generated for comments are [edit source] links on previewing a comment.
Hence why those are the only edit links we bother hiding in this patch.

Bug: T236966
Change-Id: I5f161628c27e751c85e4caee5b2db0d0c22e294e